refactor: Move class location
